fix make_all_implementors()
keeping the provisioners in all_provisioners means that the old
provisioner objects' "make_implementors" will be called again when
make_all_implementors() is called again.

as these objects have references to a cush application object, which has
a reference to the namespace that will be modified, this leads to the
same namespace object being reinitialized, which fails due to
NamespaceCollisionError's.